Convert collection update to a coroutine Add error handling to refreshable resources collection view model

+++ app/src/main/java/com/boardgamegeek/repository/GameCollectionRepository.kt
@@ -2,19 +2,18 @@ package com.boardgamegeek.repository
 
 import android.content.ContentValues
 import androidx.core.content.contentValuesOf
-import androidx.lifecycle.MutableLiveData
 import com.boardgamegeek.BggApplication
 import com.boardgamegeek.auth.AccountUtils
 import com.boardgamegeek.db.CollectionDao
 import com.boardgamegeek.entities.CollectionItemEntity
-import com.boardgamegeek.extensions.executeAsyncTask
+import com.boardgamegeek.extensions.asDateForApi
 import com.boardgamegeek.extensions.load
 import com.boardgamegeek.io.Adapter
 import com.boardgamegeek.io.BggService
 import com.boardgamegeek.mappers.CollectionItemMapper
-import com.boardgamegeek.provider.BggContract
+import com.boardgamegeek.provider.BggContract.*
+import com.boardgamegeek.provider.BggContract.Collection
 import com.boardgamegeek.service.SyncService
-import com.boardgamegeek.tasks.sync.SyncCollectionByGameTask
 import com.boardgamegeek.util.ImageUtils.getImageId
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
@@ -32,40 +31,42 @@ class GameCollectionRepository(val application: BggApplication) {
 
     suspend fun refreshCollectionItem(gameId: Int, collectionId: Int): CollectionItemEntity? =
         withContext(Dispatchers.IO) {
-            val collectionIds = mutableListOf<Int>()
+            if (gameId != INVALID_ID && !username.isNullOrBlank()) {
+                val timestamp = System.currentTimeMillis()
+                val response = Adapter.createForXmlWithAuth(application).collectionC(
+                    username, mapOf(
+                        BggService.COLLECTION_QUERY_KEY_SHOW_PRIVATE to "1",
+                        BggService.COLLECTION_QUERY_KEY_STATS to "1",
+                        BggService.COLLECTION_QUERY_KEY_ID to gameId.toString(),
+                        //BggService.COLLECTION_QUERY_KEY_SUBTYPE to subType TODO determine if needed
+                    )
+                )
 
-            val options = mutableMapOf(
-                BggService.COLLECTION_QUERY_KEY_SHOW_PRIVATE to "1",
-                BggService.COLLECTION_QUERY_KEY_STATS to "1",
-                BggService.COLLECTION_QUERY_KEY_ID to gameId.toString(),
-                //BggService.COLLECTION_QUERY_KEY_SUBTYPE to subType TODO determine if needed
-            )
-            val timestamp = System.currentTimeMillis()
-            val response = Adapter.createForXmlWithAuth(application).collectionC(username, options)
-
-            var entity: CollectionItemEntity? = null
-            response.items?.forEach { item ->
-                val pair = mapper.map(item)
-                val id = dao.saveItem(pair.first, pair.second, timestamp)
-                collectionIds.add(id.first)
-                if (pair.first.collectionId == collectionId) {
-                    entity = pair.first.copy(internalId = id.second, syncTimestamp = timestamp)
+                val collectionIds = mutableListOf<Int>()
+                var entity: CollectionItemEntity? = null
+                response.items?.forEach { collectionItem ->
+                    val (item, game) = mapper.map(collectionItem)
+                    val (id, internalId) = dao.saveItem(item, game, timestamp)
+                    collectionIds.add(id)
+                    if (item.collectionId == collectionId) {
+                        entity = item.copy(internalId = internalId, syncTimestamp = timestamp)
+                    }
                 }
-            }
-            Timber.i(
-                "Synced %,d collection item(s) for game '%s'", response.items?.size ?: 0, gameId
-            )
+                Timber.i(
+                    "Synced %,d collection item(s) for game '%s'", response.items?.size ?: 0, gameId
+                )
 
-            val deleteCount = dao.delete(gameId, collectionIds)
-            Timber.i("Removed %,d collection item(s) for game '%s'", deleteCount, gameId)
+                val deleteCount = dao.delete(gameId, collectionIds)
+                Timber.i("Removed %,d collection item(s) for game '%s'", deleteCount, gameId)
 
-            entity
+                entity
+            } else null
         }
 
     suspend fun refreshHeroImage(item: CollectionItemEntity): CollectionItemEntity = withContext(Dispatchers.IO) {
         val response = Adapter.createGeekdoApi().image2(item.thumbnailUrl.getImageId())
         val url = response.images.medium.url
-        dao.update(item.internalId, contentValuesOf(BggContract.Collection.COLLECTION_HERO_IMAGE_URL to url))
+        dao.update(item.internalId, contentValuesOf(Collection.COLLECTION_HERO_IMAGE_URL to url))
         item.copy(heroImageUrl = url)
     }
 
@@ -74,76 +75,83 @@ class GameCollectionRepository(val application: BggApplication) {
     suspend fun refreshCollectionItems(
         gameId: Int,
         subType: String = BggService.THING_SUBTYPE_BOARDGAME
-    ): List<CollectionItemEntity> = withContext(Dispatchers.IO) {
-        val list = mutableListOf<CollectionItemEntity>()
-        val mapper = CollectionItemMapper()
-        val collectionIds = arrayListOf<Int>()
-        val timestamp = System.currentTimeMillis()
-        val options = mutableMapOf(
-            BggService.COLLECTION_QUERY_KEY_SHOW_PRIVATE to "1",
-            BggService.COLLECTION_QUERY_KEY_STATS to "1",
-            BggService.COLLECTION_QUERY_KEY_ID to gameId.toString(),
-            BggService.COLLECTION_QUERY_KEY_SUBTYPE to subType
-        )
-        // TODO This doesn't sync only-played games (the played flag needs to be set explicitly)
-        val response = Adapter.createForXmlWithAuth(application).collectionC(username, options)
-        response.items?.forEach { item ->
-            val pair = mapper.map(item)
-            val collectionId = dao.saveItem(pair.first, pair.second, timestamp)
-            list += pair.first.copy(internalId = collectionId.second, syncTimestamp = timestamp)
-            collectionIds += collectionId.first
-        }
-        Timber.i("Synced %,d collection item(s) for game '%s'", response.items?.size ?: 0, gameId)
+    ): List<CollectionItemEntity>? = withContext(Dispatchers.IO) {
+        if (gameId != INVALID_ID && !username.isNullOrBlank()) {
+            val timestamp = System.currentTimeMillis()
+            val list = mutableListOf<CollectionItemEntity>()
+            // TODO This doesn't sync only-played games (the played flag needs to be set explicitly)
+            val response = Adapter.createForXmlWithAuth(application).collectionC(
+                username, mapOf(
+                    BggService.COLLECTION_QUERY_KEY_SHOW_PRIVATE to "1",
+                    BggService.COLLECTION_QUERY_KEY_STATS to "1",
+                    BggService.COLLECTION_QUERY_KEY_ID to gameId.toString(),
+                    BggService.COLLECTION_QUERY_KEY_SUBTYPE to subType
+                )
+            )
+            val collectionIds = arrayListOf<Int>()
+            response.items?.forEach { collectionItem ->
+                val (item, game) = mapper.map(collectionItem)
+                val (collectionId, internalId) = dao.saveItem(item, game, timestamp)
+                list += item.copy(internalId = internalId, syncTimestamp = timestamp)
+                collectionIds += collectionId
+            }
+            Timber.i("Synced %,d collection item(s) for game '%s'", response.items?.size ?: 0, gameId)
 
-        val deleteCount = dao.delete(gameId, collectionIds)
-        Timber.i("Removed %,d collection item(s) for game '%s'", deleteCount, gameId)
+            val deleteCount = dao.delete(gameId, collectionIds)
+            Timber.i("Removed %,d collection item(s) for game '%s'", deleteCount, gameId)
 
-        list
+            list
+        } else null
     }
 
-    fun addCollectionItem(gameId: Int, statuses: List<String>, wishListPriority: Int?) {
-        if (gameId == BggContract.INVALID_ID) return
-        application.appExecutors.diskIO.execute {
-            val values = contentValuesOf()
-            values.put(BggContract.Collection.GAME_ID, gameId)
-            putValue(statuses, values, BggContract.Collection.STATUS_OWN)
-            putValue(statuses, values, BggContract.Collection.STATUS_PREORDERED)
-            putValue(statuses, values, BggContract.Collection.STATUS_FOR_TRADE)
-            putValue(statuses, values, BggContract.Collection.STATUS_WANT)
-            putValue(statuses, values, BggContract.Collection.STATUS_WANT_TO_PLAY)
-            putValue(statuses, values, BggContract.Collection.STATUS_WANT_TO_BUY)
-            putValue(statuses, values, BggContract.Collection.STATUS_WISHLIST)
-            putValue(statuses, values, BggContract.Collection.STATUS_PREVIOUSLY_OWNED)
+    fun addCollectionItem(
+        gameId: Int,
+        statuses: List<String>,
+        wishListPriority: Int?,
+        timestamp: Long = System.currentTimeMillis()
+    ) {
+        if (gameId != INVALID_ID) {
+            val values = contentValuesOf(
+                Collection.GAME_ID to gameId,
+                Collection.STATUS_DIRTY_TIMESTAMP to timestamp
+            )
+            putValue(statuses, values, Collection.STATUS_OWN)
+            putValue(statuses, values, Collection.STATUS_PREORDERED)
+            putValue(statuses, values, Collection.STATUS_FOR_TRADE)
+            putValue(statuses, values, Collection.STATUS_WANT)
+            putValue(statuses, values, Collection.STATUS_WANT_TO_PLAY)
+            putValue(statuses, values, Collection.STATUS_WANT_TO_BUY)
+            putValue(statuses, values, Collection.STATUS_WISHLIST)
+            putValue(statuses, values, Collection.STATUS_PREVIOUSLY_OWNED)
             putWishList(statuses, wishListPriority, values)
-            values.put(BggContract.Collection.STATUS_DIRTY_TIMESTAMP, System.currentTimeMillis())
 
+            // TODO move into DAO
             application.contentResolver.load(
-                BggContract.Games.buildGameUri(gameId),
+                Games.buildGameUri(gameId),
                 arrayOf(
-                    BggContract.Games.GAME_NAME,
-                    BggContract.Games.GAME_SORT_NAME,
-                    BggContract.Games.YEAR_PUBLISHED,
-                    BggContract.Games.IMAGE_URL,
-                    BggContract.Games.THUMBNAIL_URL,
-                    BggContract.Games.HERO_IMAGE_URL
+                    Games.GAME_NAME,
+                    Games.GAME_SORT_NAME,
+                    Games.YEAR_PUBLISHED,
+                    Games.IMAGE_URL,
+                    Games.THUMBNAIL_URL,
+                    Games.HERO_IMAGE_URL
                 )
             )?.use {
                 if (it.moveToFirst()) {
-                    values.put(BggContract.Collection.COLLECTION_NAME, it.getString(0))
-                    values.put(BggContract.Collection.COLLECTION_SORT_NAME, it.getString(1))
-                    values.put(BggContract.Collection.COLLECTION_YEAR_PUBLISHED, it.getInt(2))
-                    values.put(BggContract.Collection.COLLECTION_IMAGE_URL, it.getString(3))
-                    values.put(BggContract.Collection.COLLECTION_THUMBNAIL_URL, it.getString(4))
-                    values.put(BggContract.Collection.COLLECTION_HERO_IMAGE_URL, it.getString(5))
-                    values.put(BggContract.Collection.COLLECTION_DIRTY_TIMESTAMP, System.currentTimeMillis())
+                    values.put(Collection.COLLECTION_NAME, it.getString(0))
+                    values.put(Collection.COLLECTION_SORT_NAME, it.getString(1))
+                    values.put(Collection.COLLECTION_YEAR_PUBLISHED, it.getInt(2))
+                    values.put(Collection.COLLECTION_IMAGE_URL, it.getString(3))
+                    values.put(Collection.COLLECTION_THUMBNAIL_URL, it.getString(4))
+                    values.put(Collection.COLLECTION_HERO_IMAGE_URL, it.getString(5))
+                    values.put(Collection.COLLECTION_DIRTY_TIMESTAMP, System.currentTimeMillis())
                 }
             }
 
-            val gameName = values.getAsString(BggContract.Collection.COLLECTION_NAME)
-            val response = application.contentResolver.insert(BggContract.Collection.CONTENT_URI, values)
-            val internalId = response?.lastPathSegment?.toLongOrNull()
-                ?: BggContract.INVALID_ID.toLong()
-            if (internalId == BggContract.INVALID_ID.toLong()) {
+            val gameName = values.getAsString(Collection.COLLECTION_NAME)
+            val response = application.contentResolver.insert(Collection.CONTENT_URI, values)
+            val internalId = response?.lastPathSegment?.toLongOrNull() ?: INVALID_ID.toLong()
+            if (internalId == INVALID_ID.toLong()) {
                 Timber.d("Collection item for game %s (%s) not added", gameName, gameId)
             } else {
                 Timber.d("Collection item added for game %s (%s) (internal ID = %s)", gameName, gameId, internalId)
@@ -157,43 +165,106 @@ class GameCollectionRepository(val application: BggApplication) {
     }
 
     private fun putWishList(statuses: List<String>, wishListPriority: Int?, values: ContentValues) {
-        if (statuses.contains(BggContract.Collection.STATUS_WISHLIST)) {
-            values.put(BggContract.Collection.STATUS_WISHLIST, 1)
+        if (statuses.contains(Collection.STATUS_WISHLIST)) {
+            values.put(Collection.STATUS_WISHLIST, 1)
             values.put(
-                BggContract.Collection.STATUS_WISHLIST_PRIORITY, wishListPriority ?: 3 // like to have
+                Collection.STATUS_WISHLIST_PRIORITY, wishListPriority ?: 3 // like to have
             )
-            return
         } else {
-            values.put(BggContract.Collection.STATUS_WISHLIST, 0)
+            values.put(Collection.STATUS_WISHLIST, 0)
         }
     }
 
-    fun update(internalId: Long, values: ContentValues) {
-        if (internalId == BggContract.INVALID_ID.toLong()) return
-        application.appExecutors.diskIO.execute {
-            dao.update(internalId, values)
-        }
-    }
-
-    fun resetTimestamps(internalId: Long, gameId: Int, errorMessage: MutableLiveData<String>) {
-        if (internalId == BggContract.INVALID_ID.toLong()) return
-        application.appExecutors.diskIO.execute {
+    suspend fun updatePrivateInfo(
+        internalId: Long, priceCurrency: String?,
+        price: Double?,
+        currentValueCurrency: String?,
+        currentValue: Double?,
+        quantity: Int?,
+        acquisitionDate: Long?,
+        acquiredFrom: String?,
+        inventoryLocation: String?,
+    ): Int = withContext(Dispatchers.IO) {
+        if (internalId != INVALID_ID.toLong()) {
             val values = contentValuesOf(
-                BggContract.Collection.COLLECTION_DIRTY_TIMESTAMP to 0,
-                BggContract.Collection.STATUS_DIRTY_TIMESTAMP to 0,
-                BggContract.Collection.COMMENT_DIRTY_TIMESTAMP to 0,
-                BggContract.Collection.RATING_DIRTY_TIMESTAMP to 0,
-                BggContract.Collection.PRIVATE_INFO_DIRTY_TIMESTAMP to 0,
-                BggContract.Collection.WISHLIST_COMMENT_DIRTY_TIMESTAMP to 0,
-                BggContract.Collection.TRADE_CONDITION_DIRTY_TIMESTAMP to 0,
-                BggContract.Collection.WANT_PARTS_DIRTY_TIMESTAMP to 0,
-                BggContract.Collection.HAS_PARTS_DIRTY_TIMESTAMP to 0
+                Collection.PRIVATE_INFO_DIRTY_TIMESTAMP to System.currentTimeMillis(),
+                Collection.PRIVATE_INFO_PRICE_PAID_CURRENCY to priceCurrency,
+                Collection.PRIVATE_INFO_PRICE_PAID to price,
+                Collection.PRIVATE_INFO_CURRENT_VALUE_CURRENCY to currentValueCurrency,
+                Collection.PRIVATE_INFO_CURRENT_VALUE to currentValue,
+                Collection.PRIVATE_INFO_QUANTITY to quantity,
+                Collection.PRIVATE_INFO_ACQUISITION_DATE to acquisitionDate.asDateForApi(),
+                Collection.PRIVATE_INFO_ACQUIRED_FROM to acquiredFrom,
+                Collection.PRIVATE_INFO_INVENTORY_LOCATION to inventoryLocation
             )
-
-            val success = dao.update(internalId, values) > 0
-            if (success && gameId != BggContract.INVALID_ID) {
-                SyncCollectionByGameTask(application, gameId, errorMessage).executeAsyncTask()
-            }
-        }
+            dao.update(internalId, values)
+        } else 0
     }
+
+    suspend fun updateStatuses(internalId: Long, statuses: List<String>, wishlistPriority: Int): Int =
+        withContext(Dispatchers.IO) {
+            if (internalId != INVALID_ID.toLong()) {
+                val values = contentValuesOf(
+                    Collection.STATUS_DIRTY_TIMESTAMP to System.currentTimeMillis(),
+                    Collection.STATUS_OWN to statuses.contains(Collection.STATUS_OWN),
+                    Collection.STATUS_PREVIOUSLY_OWNED to statuses.contains(Collection.STATUS_PREVIOUSLY_OWNED),
+                    Collection.STATUS_PREORDERED to statuses.contains(Collection.STATUS_PREORDERED),
+                    Collection.STATUS_FOR_TRADE to statuses.contains(Collection.STATUS_FOR_TRADE),
+                    Collection.STATUS_WANT to statuses.contains(Collection.STATUS_WANT),
+                    Collection.STATUS_WANT_TO_BUY to statuses.contains(Collection.STATUS_WANT_TO_BUY),
+                    Collection.STATUS_WANT_TO_PLAY to statuses.contains(Collection.STATUS_WANT_TO_PLAY),
+                    Collection.STATUS_WISHLIST to statuses.contains(Collection.STATUS_WISHLIST)
+                )
+                if (statuses.contains(Collection.STATUS_WISHLIST)) {
+                    values.put(Collection.STATUS_WISHLIST_PRIORITY, wishlistPriority.coerceIn(1..5))
+                }
+                dao.update(internalId, values)
+            } else 0
+        }
+
+    suspend fun updateRating(internalId: Long, rating: Double): Int = withContext(Dispatchers.IO) {
+        if (internalId != INVALID_ID.toLong()) {
+            val values = contentValuesOf(
+                Collection.RATING to rating,
+                Collection.RATING_DIRTY_TIMESTAMP to System.currentTimeMillis()
+            )
+            dao.update(internalId, values)
+        } else 0
+    }
+
+    suspend fun updateText(internalId: Long, text: String, textColumn: String, timestampColumn: String): Int =
+        withContext(Dispatchers.IO) {
+            if (internalId != INVALID_ID.toLong()) {
+                val values = contentValuesOf(
+                    textColumn to text,
+                    timestampColumn to System.currentTimeMillis()
+                )
+                dao.update(internalId, values)
+            } else 0
+        }
+
+    suspend fun markAsDeleted(internalId: Long): Int = withContext(Dispatchers.IO) {
+        if (internalId != INVALID_ID.toLong()) {
+            val values = contentValuesOf(Collection.COLLECTION_DELETE_TIMESTAMP to System.currentTimeMillis())
+            dao.update(internalId, values)
+        } else 0
+    }
+
+    suspend fun resetTimestamps(internalId: Long): Int =
+        withContext(Dispatchers.IO) {
+            if (internalId != INVALID_ID.toLong()) {
+                val values = contentValuesOf(
+                    Collection.COLLECTION_DIRTY_TIMESTAMP to 0,
+                    Collection.STATUS_DIRTY_TIMESTAMP to 0,
+                    Collection.COMMENT_DIRTY_TIMESTAMP to 0,
+                    Collection.RATING_DIRTY_TIMESTAMP to 0,
+                    Collection.PRIVATE_INFO_DIRTY_TIMESTAMP to 0,
+                    Collection.WISHLIST_COMMENT_DIRTY_TIMESTAMP to 0,
+                    Collection.TRADE_CONDITION_DIRTY_TIMESTAMP to 0,
+                    Collection.WANT_PARTS_DIRTY_TIMESTAMP to 0,
+                    Collection.HAS_PARTS_DIRTY_TIMESTAMP to 0,
+                )
+                dao.update(internalId, values)
+            } else 0
+        }
 }
